demos/mcp-playground/requirements.txt
streamlit>=1.35
openai>=1.30
tavily-python>=0.3
python-dotenv>=1.0
